ingestion/youtube/requirements.txt
kafka-python==2.0.2
google-api-python-client==2.90.0
python-dotenv==1.0.1